miden_field::word

Struct Word

#[repr(C)]
pub struct Word {
    pub a: Felt,
    pub b: Felt,
    pub c: Felt,
    pub d: Felt,
}

A unit of data consisting of 4 field elements.

For ordering a word with Ord the word’s elements are treated as limbs of an integer in little-endian limb order and thus comparison starts from the most significant element.

Fields
a: Felt

The underlying elements of this word.

b: Feltc: Feltd: Felt
Implementations

impl Word

pub const NUM_ELEMENTS: usize = 4

The number of field elements in the word.

pub const SERIALIZED_SIZE: usize = 32

The serialized size of the word in bytes.

pub const fn new(value: [Felt; 4]) -> Self

Creates a new Word from the given field elements.

pub const fn into_elements(self) -> [Felt; 4]

Returns the elements of this word as an array.

pub const fn parse(hex: &str) -> Result<Self, &'static str>

Parses a hex string into a new Word.

The input must contain valid hex prefixed with 0x. The input after the prefix must contain between 0 and 64 characters (inclusive).

The input is interpreted to have little-endian byte ordering. Nibbles are interpreted to have big-endian ordering so that “0x10” represents Felt::new(16), not Felt::new(1).

This function is usually used via the word! macro.

use miden_field::{Felt, Word, word};
let word = word!("0x1000000000000000200000000000000030000000000000004000000000000000");
assert_eq!(
    word,
    Word::new([
        Felt::new_unchecked(16),
        Felt::new_unchecked(32),
        Felt::new_unchecked(48),
        Felt::new_unchecked(64)
    ])
);

pub const fn empty() -> Self

Returns a new Word consisting of four ZERO elements.

pub fn is_empty(&self) -> bool

Returns true if the word consists of four ZERO elements.

pub fn as_elements(&self) -> &[Felt]

Returns the word as a slice of field elements.

pub fn as_bytes(&self) -> [u8; 32]

Returns the word as a byte array.

pub fn words_as_elements_iter<'a, I>(words: I) -> impl Iterator<Item = &'a Felt>
where I: Iterator<Item = &'a Self>,

Returns an iterator over the elements of multiple words.

pub fn words_as_elements(words: &[Self]) -> &[Felt]

Returns all elements of multiple words as a slice.

pub fn to_hex(&self) -> String

Returns hexadecimal representation of this word prefixed with 0x.

pub fn to_vec(&self) -> Vec<Felt>

Returns internal elements of this word as a vector.

pub fn reversed(&self) -> Self

Returns a copy of this word with its elements in reverse order.